feat:优化代码

+++ sky-take-out/sky-server/src/main/java/com/sky/mapper/DishFlavorMapper.java
@@ -29,5 +29,5 @@ public interface DishFlavorMapper {
      * @param dishId 菜品ID
      * @return
      */
-    List<DishFlavor> getByDishId(Integer dishId);
+    List<DishFlavor> getByDishId(Long dishId);
 }
